Removed Legacy Implementation and Renamed new Implementation with `r_` prefix to not haing that prefix
