+++ backend/internal/workflow/model/consignment.go
@@ -1,54 +1,90 @@
 package model
 
-import (
-	"github.com/google/uuid"
-)
+import "github.com/google/uuid"
 
-// TradeFlow represents the trade flow of a consignment.
-type TradeFlow string
+// ConsignmentFlow represents the flow type of a consignment.
+type ConsignmentFlow string
 
 const (
-	TradeFlowImport TradeFlow = "IMPORT"
-	TradeFlowExport TradeFlow = "EXPORT"
+	ConsignmentFlowImport ConsignmentFlow = "IMPORT"
+	ConsignmentFlowExport ConsignmentFlow = "EXPORT"
 )
 
-// ConsignmentState represents the state of a consignment in the workflow.
+// ConsignmentState represents the state of a consignment.
 type ConsignmentState string
 
 const (
-	ConsignmentStateInProgress     ConsignmentState = "IN_PROGRESS"
-	ConsignmentStateRequiresRework ConsignmentState = "REQUIRES_REWORK" // At least one task has been rejected
-	ConsignmentStateFinished       ConsignmentState = "FINISHED"
+	ConsignmentStateInProgress ConsignmentState = "IN_PROGRESS"
+	ConsignmentStateFinished   ConsignmentState = "FINISHED"
 )
 
-// Consignment represents the state and data of a consignment in the workflow system.
+// Consignment represents a consignment in the system.
 type Consignment struct {
 	BaseModel
-	TradeFlow     TradeFlow              `gorm:"type:varchar(20);column:trade_flow;not null" json:"tradeFlow"`                   // Type of trade flow: IMPORT or EXPORT
-	Items         []Item                 `gorm:"type:jsonb;column:items;serializer:json;not null" json:"items"`                  // List of items in the consignment
-	TraderID      string                 `gorm:"type:varchar(255);column:trader_id;not null" json:"traderId"`                    // Reference to the Trader
-	State         ConsignmentState       `gorm:"type:varchar(20);column:state;not null" json:"state"`                            // IN_PROGRESS, REQUIRES_REWORK, FINISHED
-	GlobalContext map[string]interface{} `gorm:"type:jsonb;column:global_context;serializer:json;not null" json:"globalContext"` // Global context for the consignment
+	Flow          ConsignmentFlow   `gorm:"type:varchar(50);column:flow;not null" json:"flow"`                              // e.g., IMPORT, EXPORT
+	TraderID      string            `gorm:"type:varchar(100);column:trader_id;not null" json:"traderId"`                    // ID of the trader associated with the consignment
+	State         ConsignmentState  `gorm:"type:varchar(50);column:state;not null" json:"state"`                            // State of the consignment
+	Items         []ConsignmentItem `gorm:"type:jsonb;column:items;serializer:json;not null" json:"items"`                  // Items in the consignment
+	GlobalContext map[string]any    `gorm:"type:jsonb;column:global_context;serializer:json;not null" json:"globalContext"` // Global context for the consignment
 }
 
 func (c *Consignment) TableName() string {
 	return "consignments"
 }
 
-type ConsignmentStep struct {
-	StepID    string     `json:"stepId"`    // Step ID within the workflow template
-	Type      StepType   `json:"type"`      // Type of the task
-	TaskID    uuid.UUID  `json:"taskId"`    // Associated Task ID
-	Status    TaskStatus `json:"status"`    // Current status of the task
-	DependsOn []string   `json:"dependsOn"` // List of step IDs that this step depends on
-
+// ConsignmentItem represents an individual item within a consignment.
+type ConsignmentItem struct {
+	HSCodeID     uuid.UUID `gorm:"type:uuid;column:hs_code_id;not null" json:"hsCodeId"`         // HS Code ID
+	ItemMetadata any       `gorm:"type:jsonb;column:item_metadata;not null" json:"itemMetadata"` // Metadata about the item
 }
 
-// Item represents an individual item within a consignment.
-type Item struct {
-	HSCodeID          uuid.UUID         `json:"hsCodeID"`                 // HS Code ID of the item
-	HSCode            string            `json:"hsCode"`                   // HS Code of the item
-	HSCodeDescription string            `json:"hsCodeDescription"`        // Description of the HS Code
-	AdditionalData    interface{}       `json:"additionalData,omitempty"` // Additional item data
-	Steps             []ConsignmentStep `json:"steps"`                    // List of steps associated with this item
+// CreateConsignmentItemDTO represents the data required to create a consignment item.
+type CreateConsignmentItemDTO struct {
+	HSCodeID     uuid.UUID `json:"hsCodeId" binding:"required"` // HS Code ID
+	ItemMetadata any       `json:"itemMetadata,omitempty"`      // Metadata about the item (optional)
+}
+
+// CreateConsignmentDTO represents the data required to create a consignment.
+type CreateConsignmentDTO struct {
+	Flow          ConsignmentFlow            `json:"flow" binding:"required,oneof=IMPORT EXPORT"` // e.g., IMPORT, EXPORT
+	TraderID      *string                    `json:"traderId,omitempty"`                          // ID of the trader associated with the consignment (optional)
+	Items         []CreateConsignmentItemDTO `json:"items" binding:"required,dive,required"`      // Items in the consignment
+	GlobalContext map[string]any             `json:"globalContext,omitempty"`                     // Global context for the consignment (optional)
+}
+
+// UpdateConsignmentDTO represents the data required to update a consignment.
+type UpdateConsignmentDTO struct {
+	ConsignmentID         uuid.UUID         `json:"consignmentId" binding:"required"` // Consignment ID
+	State                 *ConsignmentState `json:"state,omitempty"`                  // New state of the consignment (optional)
+	AppendToGlobalContext map[string]any    `json:"appendToGlobalContext,omitempty"`  // Additional global context to append to the consignment (optional)
+}
+
+// ConsignmentResponseDTO represents the consignment data returned in responses.
+type ConsignmentResponseDTO struct {
+	ID            uuid.UUID         `json:"id"`            // Consignment ID
+	Flow          ConsignmentFlow   `json:"flow"`          // e.g., IMPORT, EXPORT
+	TraderID      string            `json:"traderId"`      // ID of the trader associated with the consignment
+	State         ConsignmentState  `json:"state"`         // State of the consignment
+	Items         []ConsignmentItem `json:"items"`         // Items in the consignment
+	GlobalContext map[string]any    `json:"globalContext"` // Global context for the consignment
+	CreatedAt     string            `json:"createdAt"`     // Timestamp of consignment creation
+	UpdatedAt     string            `json:"updatedAt"`     // Timestamp of last consignment update
+	WorkflowNodes []WorkflowNode    `json:"workflowNodes"` // Associated workflow nodes
+}
+
+// ConsignmentListResult represents the result of querying consignments with pagination
+type ConsignmentListResult struct {
+	TotalCount int64                    `json:"totalCount"`
+	Items      []ConsignmentResponseDTO `json:"items"`
+	Offset     int                      `json:"offset"`
+	Limit      int                      `json:"limit"`
+}
+
+// ConsignmentFilter will be used when querying consignments as batch
+type ConsignmentFilter struct {
+	TraderID *string           `json:"traderId,omitempty"`
+	Flow     *ConsignmentFlow  `json:"flow,omitempty"`
+	State    *ConsignmentState `json:"state,omitempty"`
+	Offset   *int              `json:"offset,omitempty"`
+	Limit    *int              `json:"limit,omitempty"`
 }
