--- backend/internal/workflow/model/task_dtos.go
@@ -1,12 +1,0 @@
-package model
-
-import (
-	"github.com/google/uuid"
-)
-
-// TaskCompletionNotification represents a notification sent to Workflow Manager when a task completes
-type TaskCompletionNotification struct {
-	TaskID              uuid.UUID              `json:"taskId" binding:"required"`
-	State               TaskStatus             `json:"state" binding:"required"`
-	AppendGlobalContext map[string]interface{} `json:"appendGlobalContext,omitempty"` // Information to be appended to consignment global context
-}
