--- backend/internal/workflow/r_model/workflow.go
@@ -1,19 +1,0 @@
-package r_model
-
-import "github.com/google/uuid"
-
-type WorkflowTemplate struct {
-	BaseModel
-	Name          string    `gorm:"type:varchar(100);column:name;not null" json:"name"`            // Name of the workflow template
-	Description   string    `gorm:"type:text;column:description" json:"description"`               // Description of the workflow template
-	Version       string    `gorm:"type:varchar(50);column:version;not null" json:"version"`       // Version of the workflow template
-	NodeTemplates UUIDArray `gorm:"type:jsonb;column:nodes;not null;serializer:json" json:"nodes"` // Array of workflow node template IDs
-}
-
-func (wt *WorkflowTemplate) TableName() string {
-	return "workflow_templates"
-}
-
-func (wt *WorkflowTemplate) GetNodeTemplateIDs() []uuid.UUID {
-	return wt.NodeTemplates
-}
