+++ backend/internal/workflow/service/consignment_service.go
@@ -2,144 +2,51 @@ package service
 
 import (
 	"context"
-	"errors"
 	"fmt"
+	"maps"
+	"time"
 
 	"github.com/OpenNSW/nsw/internal/workflow/model"
-	"github.com/OpenNSW/nsw/utils"
 	"github.com/google/uuid"
 	"gorm.io/gorm"
 )
 
+// ConsignmentService handles consignment-related operations.
+// It coordinates between workflow templates, nodes, and the state machine.
 type ConsignmentService struct {
-	ts *TaskService
-	db *gorm.DB
+	db                          *gorm.DB
+	templateProvider            TemplateProvider
+	nodeRepo                    WorkflowNodeRepository
+	stateMachine                *WorkflowNodeStateMachine
+	preCommitValidationCallback func([]model.WorkflowNode, map[string]any) error
 }
 
-func NewConsignmentService(ts *TaskService, db *gorm.DB) *ConsignmentService {
-	return &ConsignmentService{ts: ts, db: db}
+// SetPreCommitValidationCallback sets a callback to be executed before transaction commit
+// This allows external validation (like task manager registration) to participate in the transaction
+func (s *ConsignmentService) SetPreCommitValidationCallback(callback func([]model.WorkflowNode, map[string]any) error) {
+	s.preCommitValidationCallback = callback
 }
 
-// GetAllHSCodes retrieves all HS codes from the database
-func (s *ConsignmentService) GetAllHSCodes(ctx context.Context, filter model.HSCodeFilter) (*model.HSCodeListResult, error) {
-	// Get total count first for pagination (with filter applied)
-	var totalCount int64
-	countQuery := s.db.WithContext(ctx).Model(&model.HSCode{})
-
-	// Apply the same filter to the count query
-	if filter.HSCodeStartsWith != nil && *filter.HSCodeStartsWith != "" {
-		countQuery = countQuery.Where("hs_code LIKE ?", *filter.HSCodeStartsWith+"%")
+// NewConsignmentService creates a new instance of ConsignmentService with interface dependencies.
+// This constructor allows for dependency injection and easier testing.
+func NewConsignmentService(db *gorm.DB, templateProvider TemplateProvider, nodeRepo WorkflowNodeRepository) *ConsignmentService {
+	return &ConsignmentService{
+		db:               db,
+		templateProvider: templateProvider,
+		nodeRepo:         nodeRepo,
+		stateMachine:     NewWorkflowNodeStateMachine(nodeRepo),
 	}
-
-	countResult := countQuery.Count(&totalCount)
-	if countResult.Error != nil {
-		return nil, fmt.Errorf("failed to count HS codes: %w", countResult.Error)
-	}
-
-	// If no HS codes found, return early
-	if totalCount == 0 {
-		return &model.HSCodeListResult{
-			TotalCount: 0,
-			Items:      []model.HSCode{},
-			Offset:     0,
-			Limit:      0,
-		}, nil
-	}
-
-	var hsCodes []model.HSCode
-	query := s.db.WithContext(ctx)
-
-	// Apply filter: HSCode starts with
-	if filter.HSCodeStartsWith != nil && *filter.HSCodeStartsWith != "" {
-		query = query.Where("hs_code LIKE ?", *filter.HSCodeStartsWith+"%")
-	}
-
-	// Apply pagination with defaults and limits
-	finalOffset, finalLimit := utils.GetPaginationParams(filter.Offset, filter.Limit)
-	query = query.Offset(finalOffset).Limit(finalLimit)
-
-	// Add ordering for consistent pagination
-	query = query.Order("hs_code ASC")
-
-	result := query.Find(&hsCodes)
-	if result.Error != nil {
-		return nil, fmt.Errorf("failed to retrieve HS codes: %w", result.Error)
-	}
-
-	// Prepare the result
-	hsCodeListResult := &model.HSCodeListResult{
-		TotalCount: totalCount,
-		Items:      hsCodes,
-		Offset:     finalOffset,
-		Limit:      finalLimit,
-	}
-
-	return hsCodeListResult, nil
 }
 
-// GetHSCodeByID retrieves an HS code by its ID from the database
-func (s *ConsignmentService) GetHSCodeByID(ctx context.Context, hsCodeID uuid.UUID) (*model.HSCode, error) {
-	var hsCode model.HSCode
-	result := s.db.WithContext(ctx).First(&hsCode, "id = ?", hsCodeID)
-	if result.Error != nil {
-		if errors.Is(result.Error, gorm.ErrRecordNotFound) {
-			return nil, fmt.Errorf("HS code with ID %s not found", hsCodeID)
-		}
-		return nil, fmt.Errorf("failed to retrieve HS code: %w", result.Error)
-	}
-	return &hsCode, nil
+// NewConsignmentServiceWithDefaults creates a new instance of ConsignmentService with concrete implementations.
+// This is a convenience constructor for production use.
+func NewConsignmentServiceWithDefaults(db *gorm.DB, templateService *TemplateService, workflowNodeService *WorkflowNodeService) *ConsignmentService {
+	return NewConsignmentService(db, templateService, workflowNodeService)
 }
 
-// GetWorkFlowTemplate retrieves a workflow template based on HS code and consignment type
-func (s *ConsignmentService) GetWorkFlowTemplate(ctx context.Context, hsCode *string, hsCodeID *uuid.UUID, tradeFlow model.TradeFlow) (*model.WorkflowTemplate, error) {
-	if hsCode == nil && hsCodeID == nil {
-		return nil, fmt.Errorf("either hscode or hscodeID must be provided")
-	}
-
-	query := s.db.WithContext(ctx)
-
-	if hsCodeID != nil {
-		// Query by HS code ID
-		var workflowTemplate model.WorkflowTemplate
-		err := query.
-			Joins("JOIN workflow_template_maps ON workflow_templates.id = workflow_template_maps.workflow_template_id").
-			Where("workflow_template_maps.hs_code_id = ? AND workflow_template_maps.trade_flow = ?", *hsCodeID, tradeFlow).
-			First(&workflowTemplate).Error
-
-		if err != nil {
-			if errors.Is(err, gorm.ErrRecordNotFound) {
-				return nil, fmt.Errorf("workflow template not found for HS code ID %s and trade flow %s", *hsCodeID, tradeFlow)
-			}
-			return nil, fmt.Errorf("failed to retrieve workflow template: %w", err)
-		}
-
-		return &workflowTemplate, nil
-	}
-
-	// SELECT workflow_templates.* FROM workflow_templates
-	// JOIN workflow_template_maps ON workflow_templates.id = workflow_template_maps.workflow_template_id
-	// JOIN hs_codes ON hs_codes.id = workflow_template_maps.hs_code_id
-	// WHERE hs_codes.code = ? AND workflow_template_maps.trade_flow = ?
-	var workflowTemplate model.WorkflowTemplate
-	err := query.
-		Joins("JOIN workflow_template_maps ON workflow_templates.id = workflow_template_maps.workflow_template_id").
-		Joins("JOIN hs_codes ON hs_codes.id = workflow_template_maps.hs_code_id").
-		Where("hs_codes.hs_code = ? AND workflow_template_maps.trade_flow = ?", *hsCode, tradeFlow).
-		First(&workflowTemplate).Error
-
-	if err != nil {
-		if errors.Is(err, gorm.ErrRecordNotFound) {
-			return nil, fmt.Errorf("workflow template not found for HS code %s and trade flow %s", *hsCode, tradeFlow)
-		}
-		return nil, fmt.Errorf("failed to retrieve workflow template: %w", err)
-	}
-
-	return &workflowTemplate, nil
-}
-
-// InitializeConsignment creates a new consignment with all associated tasks based on workflow templates
-// Returns the created consignment and a list of ready tasks that should be registered with Task Manager
-func (s *ConsignmentService) InitializeConsignment(ctx context.Context, createReq *model.CreateConsignmentDTO) (*model.Consignment, []*model.Task, error) {
+// InitializeConsignment initializes the consignment based on the provided creation request.
+// Returns the (created consignment response DTO and the new READY workflow nodes) or an error if the operation fails.
+func (s *ConsignmentService) InitializeConsignment(ctx context.Context, createReq *model.CreateConsignmentDTO) (*model.ConsignmentResponseDTO, []model.WorkflowNode, error) {
 	if createReq == nil {
 		return nil, nil, fmt.Errorf("create request cannot be nil")
 	}
@@ -150,436 +57,358 @@ func (s *ConsignmentService) InitializeConsignment(ctx context.Context, createRe
 		return nil, nil, fmt.Errorf("trader ID cannot be empty")
 	}
 
-	// Use a transaction to ensure atomicity
-	return s.initializeConsignmentInTx(ctx, createReq)
-}
-
-func (s *ConsignmentService) initializeConsignmentInTx(ctx context.Context, createReq *model.CreateConsignmentDTO) (*model.Consignment, []*model.Task, error) {
-	var consignment *model.Consignment
-	var readyTasks []*model.Task
-
-	err := s.db.WithContext(ctx).Transaction(func(tx *gorm.DB) error {
-		// Fetch workflow templates for all items (with transaction context and preload Steps)
-		// Always query by HSCodeID (required field) and validate against WorkflowTemplateID if provided
-		workflowTemplates := make([]*model.WorkflowTemplate, len(createReq.Items))
-		for i := range createReq.Items {
-			// Query workflow template by HS code ID and trade flow
-			workflowTemplate, err := s.GetWorkFlowTemplate(ctx, nil, &createReq.Items[i].HSCodeID, createReq.TradeFlow)
-			if err != nil {
-				return fmt.Errorf("failed to get workflow template for item %d: %w", i, err)
-			}
-			workflowTemplates[i] = workflowTemplate
-		}
-
-		// Initialize items with HSCodeID only - Steps will be populated after tasks are created
-		items := make([]model.Item, len(createReq.Items))
-		for i := range createReq.Items {
-			// fetch HSCode details from HSCodeID
-			hsCode, err := s.GetHSCodeByID(ctx, createReq.Items[i].HSCodeID)
-			if err != nil {
-				return fmt.Errorf("failed to get HS code for item %d: %w", i, err)
-			}
-
-			items[i] = model.Item{
-				HSCodeID:          createReq.Items[i].HSCodeID,
-				HSCode:            hsCode.HSCode,
-				HSCodeDescription: hsCode.Description,
-				AdditionalData:    createReq.Items[i].ItemData,
-				Steps:             []model.ConsignmentStep{},
-			}
-		}
-
-		// Prepare global context
-		globalContext := make(map[string]interface{})
-		if createReq.GlobalContext != nil {
-			globalContext = createReq.GlobalContext
-		}
-
-		consignment = &model.Consignment{
-			TradeFlow:     createReq.TradeFlow,
-			Items:         items,
-			TraderID:      *createReq.TraderID,
-			State:         model.ConsignmentStateInProgress,
-			GlobalContext: globalContext,
-		}
-
-		// Save the consignment to generate an ID
-		if err := tx.Create(consignment).Error; err != nil {
-			return fmt.Errorf("failed to create consignment: %w", err)
-		}
-
-		// Process each item in the consignment
-		for itemIdx := range consignment.Items {
-			item := &consignment.Items[itemIdx]
-			workflowTemplate := workflowTemplates[itemIdx]
-
-			// Build tasks for this item
-			tasks, err := s.buildTasksFromTemplate(consignment.ID, *workflowTemplate)
-			if err != nil {
-				return fmt.Errorf("failed to build tasks for item %d: %w", itemIdx, err)
-			}
-
-			// Save all tasks for this item using the transaction
-			taskIDs, err := s.ts.CreateTasksInTx(ctx, tx, tasks)
-			if err != nil {
-				return fmt.Errorf("failed to create tasks for item %d: %w", itemIdx, err)
-			}
-
-			// Build ConsignmentSteps from created tasks
-			steps := make([]model.ConsignmentStep, len(tasks))
-			for i, taskID := range taskIDs {
-				tasks[i].ID = taskID // Set the generated ID
-
-				// Collect ready tasks for registration with Task Manager
-				if tasks[i].Status == model.TaskStatusReady {
-					readyTasks = append(readyTasks, &tasks[i])
-				}
-
-				// Convert DependsOn map keys to slice for ConsignmentStep
-				dependsOnSlice := make([]string, 0, len(tasks[i].DependsOn))
-				for stepID := range tasks[i].DependsOn {
-					dependsOnSlice = append(dependsOnSlice, stepID)
-				}
-
-				// Create ConsignmentStep
-				steps[i] = model.ConsignmentStep{
-					StepID:    tasks[i].StepID,
-					Type:      tasks[i].Type,
-					TaskID:    taskID,
-					Status:    tasks[i].Status,
-					DependsOn: dependsOnSlice,
-				}
-			}
-
-			// Store steps in the item
-			item.Steps = steps
-		}
-
-		// Update the consignment with the populated steps
-		if err := tx.Save(consignment).Error; err != nil {
-			return fmt.Errorf("failed to update consignment with steps: %w", err)
-		}
-
-		return nil
-	})
-
+	consignment, newReadyWorkflowNodes, err := s.initializeConsignmentInTx(ctx, createReq)
 	if err != nil {
-		return nil, nil, err
+		return nil, nil, fmt.Errorf("failed to initialize consignment: %w", err)
 	}
 
-	return consignment, readyTasks, nil
+	return consignment, newReadyWorkflowNodes, nil
 }
 
-// buildTasksFromTemplate creates task instances from a workflow template
-func (s *ConsignmentService) buildTasksFromTemplate(consignmentID uuid.UUID, template model.WorkflowTemplate) ([]model.Task, error) {
-	if len(template.Steps) == 0 {
-		return nil, fmt.Errorf("workflow template has no steps")
+// initializeConsignmentInTx initializes the consignment within a transaction.
+func (s *ConsignmentService) initializeConsignmentInTx(ctx context.Context, createReq *model.CreateConsignmentDTO) (*model.ConsignmentResponseDTO, []model.WorkflowNode, error) {
+	consignment := &model.Consignment{
+		Flow:          createReq.Flow,
+		TraderID:      *createReq.TraderID,
+		State:         model.ConsignmentStateInProgress,
+		GlobalContext: createReq.GlobalContext,
 	}
 
-	tasks := make([]model.Task, 0, len(template.Steps))
-
-	for _, step := range template.Steps {
-		// Determine task status based on dependencies
-		status := model.TaskStatusReady
-		dependsOnMap := make(map[string]model.DependencyStatus)
-
-		if len(step.DependsOn) > 0 {
-			status = model.TaskStatusLocked
-			// Initialize all dependencies as INCOMPLETE
-			for _, depStepID := range step.DependsOn {
-				dependsOnMap[depStepID] = model.DependencyStatusIncomplete
-			}
+	var items []model.ConsignmentItem
+	var workflowTemplates []model.WorkflowTemplate
+	for _, itemDTO := range createReq.Items {
+		item := model.ConsignmentItem{
+			HSCodeID:     itemDTO.HSCodeID,
+			ItemMetadata: itemDTO.ItemMetadata,
 		}
-
-		// Create the task
-		task := model.Task{
-			ConsignmentID: consignmentID,
-			StepID:        step.StepID,
-			Type:          step.Type,
-			Status:        status,
-			Config:        step.Config,
-			DependsOn:     dependsOnMap,
-		}
-		tasks = append(tasks, task)
-	}
-
-	return tasks, nil
-}
-
-// UpdateTaskStatusAndPropagateChanges updates a task's status and propagates changes to dependent tasks and consignment state and return updated dependent tasks that state became READY
-func (s *ConsignmentService) UpdateTaskStatusAndPropagateChanges(ctx context.Context, taskID uuid.UUID, newStatus model.TaskStatus, consignmentGlobalContext map[string]interface{}) ([]*model.Task, *model.Consignment, error) {
-	if taskID == uuid.Nil {
-		return nil, nil, fmt.Errorf("task ID cannot be nil")
-	}
-	if newStatus == "" {
-		return nil, nil, fmt.Errorf("task status cannot be empty")
-	}
-
-	var newReadyStateDependentTasks []*model.Task
-	var consignment *model.Consignment
-
-	err := s.db.WithContext(ctx).Transaction(func(tx *gorm.DB) error {
-		// Retrieve the task to be updated
-		task, err := s.ts.GetTaskByIDInTx(ctx, tx, taskID)
+		items = append(items, item)
+		workflowTemplate, err := s.templateProvider.GetWorkflowTemplateByHSCodeIDAndFlow(ctx, itemDTO.HSCodeID, createReq.Flow)
 		if err != nil {
-			return err
+			return nil, nil, fmt.Errorf("failed to get workflow template for HS code %s and flow %s: %w", itemDTO.HSCodeID, createReq.Flow, err)
 		}
+		workflowTemplates = append(workflowTemplates, *workflowTemplate)
+	}
+	consignment.Items = items
 
-		if task.Status == newStatus && consignmentGlobalContext == nil {
-			// No change in status and no global context update, nothing to do
-			return nil
+	// Initiate Transaction
+	tx := s.db.WithContext(ctx).Begin()
+	defer func() {
+		if r := recover(); r != nil {
+			tx.Rollback()
 		}
+	}()
 
-		// If no change in task status but global context needs to be updated
-		if task.Status == newStatus && consignmentGlobalContext != nil {
-			// Only update the consignment with the global context
-			consignment, err = s.updateConsignment(ctx, tx, task.ConsignmentID, nil, nil, &consignmentGlobalContext)
-			if err != nil {
-				return fmt.Errorf("failed to update consignment global context: %w", err)
-			}
-			return nil
-		}
+	// Create Consignment
+	if err := tx.Create(consignment).Error; err != nil {
+		tx.Rollback()
+		return nil, nil, fmt.Errorf("failed to create consignment: %w", err)
+	}
 
-		// Update the task status
-		task.Status = newStatus
-		if err := s.ts.UpdateTaskInTx(ctx, tx, task); err != nil {
-			return fmt.Errorf("failed to update task status: %w", err)
-		}
-
-		// Update dependent tasks
-		readyDependentTasks, stepsToUpdate, consignmentState, err := s.updateDependentTasks(ctx, tx, *task)
-		if err != nil {
-			return fmt.Errorf("failed to update dependent tasks: %w", err)
-		}
-
-		// Add current task's status update to the steps to update
-		stepsToUpdate = append(stepsToUpdate, &model.StepStatusUpdateDTO{
-			StepID:    task.StepID,
-			NewStatus: task.Status,
-		})
-
-		// Update the consignment state, steps, and global context
-		consignment, err = s.updateConsignment(ctx, tx, task.ConsignmentID, consignmentState, stepsToUpdate, &consignmentGlobalContext)
-		if err != nil {
-			return fmt.Errorf("failed to update consignment: %w", err)
-		}
-
-		// Return the newly READY dependent tasks
-		newReadyStateDependentTasks = readyDependentTasks
-		return nil
-	})
-
+	// Create Workflow Nodes
+	workflowNodes, newReadyWorkflowNodes, err := s.createWorkflowNodesInTx(ctx, tx, consignment.ID, workflowTemplates)
 	if err != nil {
-		return nil, nil, err
+		tx.Rollback()
+		return nil, nil, fmt.Errorf("failed to create workflow nodes: %w", err)
 	}
 
-	return newReadyStateDependentTasks, consignment, nil
+	// Execute pre-commit validation callback if set (e.g., task manager registration)
+	// This ensures external dependencies are validated before committing the transaction
+	if s.preCommitValidationCallback != nil && len(newReadyWorkflowNodes) > 0 {
+		if err := s.preCommitValidationCallback(newReadyWorkflowNodes, consignment.GlobalContext); err != nil {
+			tx.Rollback()
+			return nil, nil, fmt.Errorf("pre-commit validation failed: %w", err)
+		}
+	}
+
+	// Commit Transaction
+	if err := tx.Commit().Error; err != nil {
+		tx.Rollback()
+		return nil, nil, fmt.Errorf("failed to commit transaction: %w", err)
+	}
+
+	// Prepare Response DTO
+	responseDTO := &model.ConsignmentResponseDTO{
+		ID:            consignment.ID,
+		Flow:          consignment.Flow,
+		TraderID:      consignment.TraderID,
+		State:         consignment.State,
+		Items:         consignment.Items,
+		GlobalContext: consignment.GlobalContext,
+		CreatedAt:     consignment.CreatedAt.Format(time.RFC3339),
+		UpdatedAt:     consignment.UpdatedAt.Format(time.RFC3339),
+		WorkflowNodes: workflowNodes,
+	}
+
+	return responseDTO, newReadyWorkflowNodes, nil
 }
 
-// updateDependentTasks updates tasks that depend on the completed task and returns newly READY tasks and whether consignment is completed or steps to update
-func (s *ConsignmentService) updateDependentTasks(ctx context.Context, tx *gorm.DB, completedTask model.Task) ([]*model.Task, []*model.StepStatusUpdateDTO, *model.ConsignmentState, error) {
-	// If the completed task has been marked as REJECTED, need to update consignment state to REQUIRES_REWORK
-	if completedTask.Status == model.TaskStatusRejected {
-		consignmentStatus := model.ConsignmentStateRequiresRework
-		return []*model.Task{}, []*model.StepStatusUpdateDTO{}, &consignmentStatus, nil
-	}
-	// If the completed task is not marked as COMPLETED, no need to update dependents
-	if completedTask.Status != model.TaskStatusCompleted {
-		return []*model.Task{}, []*model.StepStatusUpdateDTO{}, nil, nil
+// createWorkflowNodesInTx builds workflow nodes for the consignment within a transaction.
+func (s *ConsignmentService) createWorkflowNodesInTx(ctx context.Context, tx *gorm.DB, consignmentID uuid.UUID, workflowTemplates []model.WorkflowTemplate) ([]model.WorkflowNode, []model.WorkflowNode, error) {
+	// Collect unique node template IDs from all workflow templates
+	uniqueNodeTemplateIDs := make(map[uuid.UUID]bool)
+	for _, wt := range workflowTemplates {
+		nodeTemplateIDs := wt.GetNodeTemplateIDs()
+		for _, nodeTemplateID := range nodeTemplateIDs {
+			uniqueNodeTemplateIDs[nodeTemplateID] = true
+		}
 	}
 
-	// Get all tasks in the same consignment
-	allTasks, err := s.ts.GetTasksByConsignmentID(ctx, completedTask.ConsignmentID)
+	// Fetch all node templates in a single query
+	nodeTemplateIDsList := make([]uuid.UUID, 0, len(uniqueNodeTemplateIDs))
+	for id := range uniqueNodeTemplateIDs {
+		nodeTemplateIDsList = append(nodeTemplateIDsList, id)
+	}
+
+	nodeTemplates, err := s.templateProvider.GetWorkflowNodeTemplatesByIDs(ctx, nodeTemplateIDsList)
 	if err != nil {
-		return nil, nil, nil, fmt.Errorf("failed to retrieve tasks for consignment %s: %w", completedTask.ConsignmentID, err)
+		return nil, nil, fmt.Errorf("failed to retrieve workflow node templates: %w", err)
 	}
 
-	// Collect tasks that need updates for batch processing
-	tasksToUpdate := make([]*model.Task, 0)
-
-	// Collect the steps that need to be updated in the consignment
-	stepsToUpdate := make([]*model.StepStatusUpdateDTO, 0)
-
-	// Collect the dependent tasks that became READY
-	var readyDependentTasks []*model.Task
-
-	// Variable to track if consignment state needs to be updated
-	isAllCompleted := true
-
-	// Find tasks that depend on the completed task
-	for i := range allTasks {
-		dependentTask := &allTasks[i]
-
-		// Check if this task depends on the completed task
-		if _, exists := dependentTask.DependsOn[completedTask.StepID]; exists {
-			// Mark this dependency as completed
-			dependentTask.DependsOn[completedTask.StepID] = model.DependencyStatusCompleted
-
-			// Check if all dependencies are now completed
-			allDepsCompleted := true
-			for _, status := range dependentTask.DependsOn {
-				if status == model.DependencyStatusIncomplete {
-					allDepsCompleted = false
-					break
-				}
-			}
-
-			// If all dependencies are completed and task was locked, make it ready
-			if allDepsCompleted && dependentTask.Status == model.TaskStatusLocked {
-				dependentTask.Status = model.TaskStatusReady
-				readyDependentTasks = append(readyDependentTasks, dependentTask)
-			}
-
-			tasksToUpdate = append(tasksToUpdate, dependentTask)
-			stepsToUpdate = append(stepsToUpdate, &model.StepStatusUpdateDTO{
-				StepID:    dependentTask.StepID,
-				NewStatus: dependentTask.Status,
-			})
-		}
-		if dependentTask.Status != model.TaskStatusCompleted && dependentTask.ID != completedTask.ID {
-			isAllCompleted = false
-		}
-	}
-
-	// Batch update all modified tasks using TaskService
-	if len(tasksToUpdate) > 0 {
-		if err := s.ts.UpdateTasksInTx(ctx, tx, tasksToUpdate); err != nil {
-			return nil, nil, nil, fmt.Errorf("failed to update dependent tasks: %w", err)
-		}
-	}
-
-	consignmentStatus := model.ConsignmentStateInProgress
-	// If all tasks in the consignment are completed, update consignment state to FINISHED
-	if isAllCompleted {
-		consignmentStatus = model.ConsignmentStateFinished
-	}
-
-	return readyDependentTasks, stepsToUpdate, &consignmentStatus, nil
+	// Delegate to the state machine for node initialization
+	return s.stateMachine.InitializeNodesFromTemplates(ctx, tx, consignmentID, nodeTemplates)
 }
 
-// getConsignmentByID retrieves a consignment by its ID from the database
-func (s *ConsignmentService) getConsignmentByID(ctx context.Context, consignmentID uuid.UUID) (*model.Consignment, error) {
+// GetConsignmentByID retrieves a consignment by its ID from the database.
+func (s *ConsignmentService) GetConsignmentByID(ctx context.Context, consignmentID uuid.UUID) (*model.ConsignmentResponseDTO, error) {
 	var consignment model.Consignment
 	result := s.db.WithContext(ctx).First(&consignment, "id = ?", consignmentID)
 	if result.Error != nil {
-		if errors.Is(result.Error, gorm.ErrRecordNotFound) {
-			return nil, fmt.Errorf("consignment %s not found", consignmentID)
-		}
-		return nil, fmt.Errorf("failed to retrieve consignment: %w", result.Error)
-	}
-	return &consignment, nil
-}
-
-// updateConsignment updates the consignment state, steps status, and global context in a transaction
-func (s *ConsignmentService) updateConsignment(ctx context.Context, tx *gorm.DB, consignmentID uuid.UUID, newState *model.ConsignmentState, stepsToUpdate []*model.StepStatusUpdateDTO, appendGlobalContext *map[string]interface{}) (*model.Consignment, error) {
-	var consignment model.Consignment
-	if err := tx.WithContext(ctx).First(&consignment, "id = ?", consignmentID).Error; err != nil {
-		return nil, fmt.Errorf("failed to retrieve consignment %s: %w", consignmentID, err)
+		return nil, fmt.Errorf("failed to retrieve consignment with ID %s: %w", consignmentID, result.Error)
 	}
 
-	// Update consignment state if provided
-	if newState != nil {
-		consignment.State = *newState
-	}
-	// Append to global context if provided
-	if appendGlobalContext != nil {
-		for k, v := range *appendGlobalContext {
-			// TODO: Need to change back to erroring out if key already exists
-			if _, exists := consignment.GlobalContext[k]; !exists {
-				consignment.GlobalContext[k] = v
-			}
-		}
-	}
-
-	// Update steps if any
-	if len(stepsToUpdate) > 0 {
-		stepMap := make(map[string]*model.ConsignmentStep)
-		for i := range consignment.Items {
-			item := &consignment.Items[i]
-			for j := range item.Steps {
-				step := &item.Steps[j]
-				stepMap[step.StepID] = step
-			}
-		}
-
-		for _, stepUpdate := range stepsToUpdate {
-			if step, ok := stepMap[stepUpdate.StepID]; ok {
-				step.Status = stepUpdate.NewStatus
-			}
-		}
-	}
-
-	// Save the updated consignment
-	if err := tx.WithContext(ctx).Save(&consignment).Error; err != nil {
-		return nil, fmt.Errorf("failed to update consignment %s: %w", consignmentID, err)
-	}
-
-	return &consignment, nil
-}
-
-// GetConsignmentByID retrieves a consignment by its ID.
-func (s *ConsignmentService) GetConsignmentByID(ctx context.Context, consignmentID uuid.UUID) (*model.ConsignmentResponse, error) {
-	if consignmentID == uuid.Nil {
-		return nil, fmt.Errorf("consignment ID cannot be nil")
-	}
-
-	consignment, err := s.getConsignmentByID(ctx, consignmentID)
+	// Retrieve associated workflow nodes
+	workflowNodes, err := s.nodeRepo.GetWorkflowNodesByConsignmentIDInTx(ctx, s.db, consignment.ID)
 	if err != nil {
-		return nil, err
+		return nil, fmt.Errorf("failed to retrieve workflow nodes for consignment %s: %w", consignmentID, err)
 	}
 
-	response := consignment.ToConsignmentResponse()
-	return &response, nil
+	responseDTO := &model.ConsignmentResponseDTO{
+		ID:            consignment.ID,
+		Flow:          consignment.Flow,
+		TraderID:      consignment.TraderID,
+		State:         consignment.State,
+		Items:         consignment.Items,
+		GlobalContext: consignment.GlobalContext,
+		CreatedAt:     consignment.CreatedAt.Format(time.RFC3339),
+		UpdatedAt:     consignment.UpdatedAt.Format(time.RFC3339),
+		WorkflowNodes: workflowNodes,
+	}
+
+	return responseDTO, nil
 }
 
-// GetConsignmentsByTraderID retrieves all consignments for a specific trader.
-func (s *ConsignmentService) GetConsignmentsByTraderID(ctx context.Context, filter model.ConsignmentTraderFilter) (model.ConsignmentListResponseDTO, error) {
-	if filter.TraderID == "" {
-		return model.ConsignmentListResponseDTO{}, fmt.Errorf("trader ID cannot be empty")
-	}
-
-	// Get total count first for pagination
-	var totalCount int64
-	countResult := s.db.WithContext(ctx).Model(&model.Consignment{}).Where("trader_id = ?", filter.TraderID).Count(&totalCount)
-	if countResult.Error != nil {
-		return model.ConsignmentListResponseDTO{}, fmt.Errorf("failed to count consignments: %w", countResult.Error)
-	}
-
-	// If no consignments found, return early
-	if totalCount == 0 {
-		return model.ConsignmentListResponseDTO{
-			TotalCount: 0,
-			Items:      []model.ConsignmentResponse{},
-			Offset:     0,
-			Limit:      0,
-		}, nil
-	}
-
+// GetConsignmentsByTraderID retrieves consignments associated with a specific trader ID.
+func (s *ConsignmentService) GetConsignmentsByTraderID(ctx context.Context, traderID string) ([]model.ConsignmentResponseDTO, error) {
 	var consignments []model.Consignment
-	query := s.db.WithContext(ctx).Where("trader_id = ?", filter.TraderID).Order("created_at DESC")
-
-	// Apply pagination with defaults and limits
-	finalOffset, finalLimit := utils.GetPaginationParams(filter.Offset, filter.Limit)
-	query = query.Offset(finalOffset).Limit(finalLimit)
-
-	result := query.Find(&consignments)
+	result := s.db.WithContext(ctx).Where("trader_id = ?", traderID).Find(&consignments)
 	if result.Error != nil {
-		return model.ConsignmentListResponseDTO{}, fmt.Errorf("failed to retrieve consignments: %w", result.Error)
+		return nil, fmt.Errorf("failed to retrieve consignments for trader %s: %w", traderID, result.Error)
 	}
 
-	// Convert to response DTOs
-	responseItems := make([]model.ConsignmentResponse, len(consignments))
-	for i, consignment := range consignments {
-		responseItems[i] = consignment.ToConsignmentResponse()
+	if len(consignments) == 0 {
+		return []model.ConsignmentResponseDTO{}, nil
 	}
 
-	// Prepare the result
-	consignmentListResult := model.ConsignmentListResponseDTO{
-		TotalCount: totalCount,
-		Items:      responseItems,
-		Offset:     finalOffset,
-		Limit:      finalLimit,
+	// Extract all consignment IDs for batch query
+	consignmentIDs := make([]uuid.UUID, len(consignments))
+	for i, c := range consignments {
+		consignmentIDs[i] = c.ID
 	}
 
-	return consignmentListResult, nil
+	// Fetch all workflow nodes in a single query to avoid N+1
+	allWorkflowNodes, err := s.nodeRepo.GetWorkflowNodesByConsignmentIDsInTx(ctx, s.db, consignmentIDs)
+	if err != nil {
+		return nil, fmt.Errorf("failed to retrieve workflow nodes for trader %s consignments: %w", traderID, err)
+	}
+
+	// Group workflow nodes by consignment ID for efficient lookup
+	nodesByConsignment := make(map[uuid.UUID][]model.WorkflowNode)
+	for _, node := range allWorkflowNodes {
+		nodesByConsignment[node.ConsignmentID] = append(nodesByConsignment[node.ConsignmentID], node)
+	}
+
+	// Build DTOs
+	var consignmentDTOs []model.ConsignmentResponseDTO
+	for _, consignment := range consignments {
+		dto := model.ConsignmentResponseDTO{
+			ID:            consignment.ID,
+			Flow:          consignment.Flow,
+			TraderID:      consignment.TraderID,
+			State:         consignment.State,
+			Items:         consignment.Items,
+			GlobalContext: consignment.GlobalContext,
+			CreatedAt:     consignment.CreatedAt.Format(time.RFC3339),
+			UpdatedAt:     consignment.UpdatedAt.Format(time.RFC3339),
+			WorkflowNodes: nodesByConsignment[consignment.ID],
+		}
+		consignmentDTOs = append(consignmentDTOs, dto)
+	}
+
+	return consignmentDTOs, nil
+}
+
+// UpdateConsignment updates an existing consignment in the database.
+func (s *ConsignmentService) UpdateConsignment(ctx context.Context, updateReq *model.UpdateConsignmentDTO) (*model.ConsignmentResponseDTO, error) {
+	if updateReq == nil {
+		return nil, fmt.Errorf("update request cannot be nil")
+	}
+
+	var consignment model.Consignment
+	result := s.db.WithContext(ctx).First(&consignment, "id = ?", updateReq.ConsignmentID)
+	if result.Error != nil {
+		return nil, fmt.Errorf("failed to retrieve consignment with ID %s for update: %w", updateReq.ConsignmentID, result.Error)
+	}
+
+	// Apply updates
+	if updateReq.State != nil {
+		consignment.State = *updateReq.State
+	}
+	if updateReq.AppendToGlobalContext != nil {
+		if consignment.GlobalContext == nil {
+			consignment.GlobalContext = make(map[string]any)
+		}
+		maps.Copy(consignment.GlobalContext, updateReq.AppendToGlobalContext)
+		// TODO: Implement the global context key selection such that no overwriting occurs.
+	}
+
+	// Save updates
+	saveResult := s.db.WithContext(ctx).Save(&consignment)
+	if saveResult.Error != nil {
+		return nil, fmt.Errorf("failed to update consignment: %w", saveResult.Error)
+	}
+
+	// Retrieve associated workflow nodes
+	workflowNodes, err := s.nodeRepo.GetWorkflowNodesByConsignmentIDInTx(ctx, s.db, consignment.ID)
+	if err != nil {
+		return nil, fmt.Errorf("failed to retrieve workflow nodes for consignment %s: %w", updateReq.ConsignmentID, err)
+	}
+
+	// Prepare response DTO
+	responseDTO := &model.ConsignmentResponseDTO{
+		ID:            consignment.ID,
+		Flow:          consignment.Flow,
+		TraderID:      consignment.TraderID,
+		State:         consignment.State,
+		Items:         consignment.Items,
+		GlobalContext: consignment.GlobalContext,
+		CreatedAt:     consignment.CreatedAt.Format(time.RFC3339),
+		UpdatedAt:     consignment.UpdatedAt.Format(time.RFC3339),
+		WorkflowNodes: workflowNodes,
+	}
+
+	return responseDTO, nil
+}
+
+// UpdateWorkflowNodeStateAndPropagateChanges updates the state of a workflow node and propagates changes to dependent nodes and consignment state, returns the new READY nodes and newGlobalContext.
+func (s *ConsignmentService) UpdateWorkflowNodeStateAndPropagateChanges(ctx context.Context, updateReq *model.UpdateWorkflowNodeDTO) ([]model.WorkflowNode, map[string]any, error) {
+	if updateReq == nil {
+		return nil, nil, fmt.Errorf("update request cannot be nil")
+	}
+
+	// Start a transaction
+	tx := s.db.WithContext(ctx).Begin()
+	defer func() {
+		if r := recover(); r != nil {
+			tx.Rollback()
+		}
+	}()
+
+	// Update the workflow node state and propagate changes, getting new READY nodes
+	newReadyNodes, newGlobalContext, err := s.updateWorkflowNodeStateAndPropagateChangesInTx(ctx, tx, updateReq)
+	if err != nil {
+		tx.Rollback()
+		return nil, nil, fmt.Errorf("failed to update workflow node state and propagate changes: %w", err)
+	}
+
+	// Commit the transaction
+	if err := tx.Commit().Error; err != nil {
+		tx.Rollback()
+		return nil, nil, fmt.Errorf("failed to commit transaction: %w", err)
+	}
+
+	return newReadyNodes, newGlobalContext, nil
+}
+
+// updateWorkflowNodeStateAndPropagateChangesInTx updates the workflow node state and propagates changes within a transaction, and returns the new READY nodes.
+func (s *ConsignmentService) updateWorkflowNodeStateAndPropagateChangesInTx(ctx context.Context, tx *gorm.DB, updateReq *model.UpdateWorkflowNodeDTO) ([]model.WorkflowNode, map[string]any, error) {
+	// Get the workflow node
+	workflowNode, err := s.nodeRepo.GetWorkflowNodeByIDInTx(ctx, tx, updateReq.WorkflowNodeID)
+	if err != nil {
+		return nil, nil, fmt.Errorf("failed to retrieve workflow node with ID %s: %w", updateReq.WorkflowNodeID, err)
+	}
+
+	var newReadyNodes []model.WorkflowNode
+
+	// Handle state transitions using the state machine
+	switch updateReq.State {
+	case model.WorkflowNodeStateFailed:
+		if workflowNode.State != model.WorkflowNodeStateFailed {
+			if err := s.stateMachine.TransitionToFailed(ctx, tx, workflowNode); err != nil {
+				return nil, nil, fmt.Errorf("failed to transition node to FAILED: %w", err)
+			}
+		}
+
+	case model.WorkflowNodeStateCompleted:
+		if workflowNode.State != model.WorkflowNodeStateCompleted {
+			result, err := s.stateMachine.TransitionToCompleted(ctx, tx, workflowNode)
+			if err != nil {
+				return nil, nil, fmt.Errorf("failed to transition node to COMPLETED: %w", err)
+			}
+			newReadyNodes = result.NewReadyNodes
+
+			// Update consignment state if all nodes are completed
+			if result.AllNodesCompleted {
+				if err := s.markConsignmentAsFinished(ctx, tx, workflowNode.ConsignmentID); err != nil {
+					return nil, nil, err
+				}
+			}
+		}
+	}
+
+	// Handle global context updates
+	var globalContext map[string]any
+	if len(updateReq.AppendGlobalContext) > 0 {
+		globalContext, err = s.appendToConsignmentGlobalContext(ctx, tx, workflowNode.ConsignmentID, updateReq.AppendGlobalContext)
+		if err != nil {
+			return nil, nil, err
+		}
+	}
+
+	return newReadyNodes, globalContext, nil
+}
+
+// markConsignmentAsFinished updates the consignment state to FINISHED.
+func (s *ConsignmentService) markConsignmentAsFinished(ctx context.Context, tx *gorm.DB, consignmentID uuid.UUID) error {
+	var consignment model.Consignment
+	result := tx.WithContext(ctx).First(&consignment, "id = ?", consignmentID)
+	if result.Error != nil {
+		return fmt.Errorf("failed to retrieve consignment %s: %w", consignmentID, result.Error)
+	}
+
+	consignment.State = model.ConsignmentStateFinished
+	if err := tx.WithContext(ctx).Save(&consignment).Error; err != nil {
+		return fmt.Errorf("failed to update consignment %s state to FINISHED: %w", consignmentID, err)
+	}
+
+	return nil
+}
+
+// appendToConsignmentGlobalContext appends key-value pairs to the consignment's global context.
+func (s *ConsignmentService) appendToConsignmentGlobalContext(ctx context.Context, tx *gorm.DB, consignmentID uuid.UUID, appendContext map[string]any) (map[string]any, error) {
+	var consignment model.Consignment
+	result := tx.WithContext(ctx).First(&consignment, "id = ?", consignmentID)
+	if result.Error != nil {
+		return nil, fmt.Errorf("failed to retrieve consignment %s: %w", consignmentID, result.Error)
+	}
+
+	if consignment.GlobalContext == nil {
+		consignment.GlobalContext = make(map[string]any)
+	}
+	maps.Copy(consignment.GlobalContext, appendContext)
+
+	if err := tx.WithContext(ctx).Save(&consignment).Error; err != nil {
+		return nil, fmt.Errorf("failed to update consignment %s global context: %w", consignmentID, err)
+	}
+
+	return consignment.GlobalContext, nil
 }
