+++ backend/internal/workflow/service/template_service.go
@@ -1,9 +1,9 @@
-package r_service
+package service
 
 import (
 	"context"
 
-	"github.com/OpenNSW/nsw/internal/workflow/r_model"
+	"github.com/OpenNSW/nsw/internal/workflow/model"
 	"github.com/google/uuid"
 	"gorm.io/gorm"
 )
@@ -20,8 +20,8 @@ func NewTemplateService(db *gorm.DB) *TemplateService {
 }
 
 // GetWorkflowTemplateByHSCodeIDAndFlow retrieves the workflow template associated with a given HS code and consignment flow.
-func (s *TemplateService) GetWorkflowTemplateByHSCodeIDAndFlow(ctx context.Context, hsCodeID uuid.UUID, flow r_model.ConsignmentFlow) (*r_model.WorkflowTemplate, error) {
-	var workflowTemplate r_model.WorkflowTemplate
+func (s *TemplateService) GetWorkflowTemplateByHSCodeIDAndFlow(ctx context.Context, hsCodeID uuid.UUID, flow model.ConsignmentFlow) (*model.WorkflowTemplate, error) {
+	var workflowTemplate model.WorkflowTemplate
 	result := s.db.WithContext(ctx).Table("workflow_templates").
 		Select("workflow_templates.*").
 		Joins("JOIN workflow_template_maps ON workflow_templates.id = workflow_template_maps.workflow_template_id").
@@ -35,8 +35,8 @@ func (s *TemplateService) GetWorkflowTemplateByHSCodeIDAndFlow(ctx context.Conte
 }
 
 // GetWorkflowNodeTemplatesByIDs retrieves workflow node templates by their IDs.
-func (s *TemplateService) GetWorkflowNodeTemplatesByIDs(ctx context.Context, ids []uuid.UUID) ([]r_model.WorkflowNodeTemplate, error) {
-	var templates []r_model.WorkflowNodeTemplate
+func (s *TemplateService) GetWorkflowNodeTemplatesByIDs(ctx context.Context, ids []uuid.UUID) ([]model.WorkflowNodeTemplate, error) {
+	var templates []model.WorkflowNodeTemplate
 	result := s.db.WithContext(ctx).Where("id IN ?", ids).Find(&templates)
 	if result.Error != nil {
 		return nil, result.Error
@@ -45,8 +45,8 @@ func (s *TemplateService) GetWorkflowNodeTemplatesByIDs(ctx context.Context, ids
 }
 
 // GetWorkflowNodeTemplateByID retrieves a workflow node template by its ID.
-func (s *TemplateService) GetWorkflowNodeTemplateByID(ctx context.Context, id uuid.UUID) (*r_model.WorkflowNodeTemplate, error) {
-	var template r_model.WorkflowNodeTemplate
+func (s *TemplateService) GetWorkflowNodeTemplateByID(ctx context.Context, id uuid.UUID) (*model.WorkflowNodeTemplate, error) {
+	var template model.WorkflowNodeTemplate
 	result := s.db.WithContext(ctx).First(&template, "id = ?", id)
 	if result.Error != nil {
 		return nil, result.Error
